requirements.txt
asgiref==3.6.0
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.4
django-filter==22.1
djangorestframework==3.14.0
drf-nested-routers==0.93.4
idna==3.4
mysqlclient==2.1.1
pycodestyle==2.10.0
pytz==2022.7
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.13
